# packages/contract/openapi.yaml
openapi: 3.1.0
info:
  title: PathPulse Backend Core API
  version: 0.1.0
  description: >
    Shared API consumed by the Android, iOS and Web clients. Phase 1 (D1) surface:
    health, managed accounts, Google + SEP-10 wallet auth, delegated
    transaction build/submit.
    Later phases add payouts, off-ramp, settlement and SCOUT endpoints.
servers:
  - url: http://localhost:8080
    description: Local dev (testnet)
paths:
  /health:
    get:
      summary: Liveness + network info
      operationId: getHealth
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HealthResponse' }
  /v1/accounts/distribution:
    get:
      summary: List protocol-governed distribution accounts
      operationId: listDistributionAccounts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DistributionAccount' }
  /v1/treasury/config:
    get:
      summary: Current multisig treasury configuration (testnet)
      operationId: getTreasuryConfig
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TreasuryConfig' }
  /v1/auth/google/verify:
    post:
      summary: Verify a Google ID token and start a session (custodial)
      operationId: verifyGoogleIdToken
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GoogleVerifyRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GoogleVerifyResponse' }
  /v1/auth/challenge:
    get:
      summary: Get a SEP-10 challenge transaction for a Stellar account
      operationId: getWalletChallenge
      parameters:
        - name: account
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WalletChallengeResponse' }
  /v1/auth/wallet/verify:
    post:
      summary: Verify a signed SEP-10 challenge and start a session (non-custodial)
      operationId: verifyWalletChallenge
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WalletVerifyRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WalletVerifyResponse' }
  /v1/auth/me:
    get:
      summary: Get the current session user, if any
      operationId: getSessionUser
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthMeResponse' }
  /v1/auth/logout:
    post:
      summary: Clear the session cookie
      operationId: logout
      responses:
        '200':
          description: OK
  /v1/tx/build:
    post:
      summary: Build (and, for managed accounts, delegate-sign) a transaction
      operationId: buildTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BuildTransactionRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BuildTransactionResponse' }
  /v1/tx/submit:
    post:
      summary: Submit a signed transaction envelope to Horizon
      operationId: submitTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubmitTransactionRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubmitTransactionResponse' }
  /v1/settlement/batches:
    get:
      summary: List settlement batches (indexer)
      operationId: listSettlementBatches
      parameters:
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, maximum: 100 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SettlementBatchPage' }
    post:
      summary: Execute a deterministic 50/30/20 settlement batch
      operationId: createSettlementBatch
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSettlementBatchRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SettlementBatch' }
  /v1/settlement/batches/{id}:
    get:
      summary: Settlement batch drill-down (Treasury → Split → Driver)
      operationId: getSettlementBatch
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SettlementBatch' }
  /v1/offramp/sessions:
    get:
      summary: List fiat off-ramp withdrawal sessions
      operationId: listOffRampSessions
      parameters:
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, maximum: 100 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OffRampSessionPage' }
    post:
      summary: Start a Mercuryo SEP-24 interactive withdrawal
      operationId: createOffRampWithdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOffRampWithdrawalRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OffRampSession' }
  /v1/offramp/sessions/{id}:
    get:
      summary: Off-ramp session status (polled)
      operationId: getOffRampSession
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OffRampSession' }
components:
  schemas:
    HealthResponse:
      type: object
      required: [status, network, horizon, version]
      properties:
        status: { type: string, enum: [ok] }
        network: { type: string, enum: [testnet, mainnet] }
        horizon: { type: string }
        version: { type: string }
    DistributionAccount:
      type: object
      required: [role, publicKey, multisig, network]
      properties:
        role: { type: string, enum: [partner_revenue, driver_pool, treasury] }
        publicKey: { type: string }
        multisig: { type: boolean }
        network: { type: string, enum: [testnet, mainnet] }
    TreasuryConfig:
      type: object
      required: [publicKey, signers, thresholds, network]
      properties:
        publicKey: { type: string }
        signers:
          type: array
          items:
            type: object
            properties:
              publicKey: { type: string }
              weight: { type: integer }
        thresholds:
          type: object
          properties:
            low: { type: integer }
            medium: { type: integer }
            high: { type: integer }
        network: { type: string, enum: [testnet, mainnet] }
    SessionUser:
      type: object
      required: [userId, method]
      properties:
        userId: { type: string }
        method: { type: string, enum: [google, wallet] }
        email: { type: string }
        address: { type: string }
    AuthMeResponse:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - { $ref: '#/components/schemas/SessionUser' }
            - type: 'null'
    GoogleVerifyRequest:
      type: object
      required: [idToken]
      properties:
        idToken: { type: string }
    GoogleVerifyResponse:
      type: object
      required: [userId, wallet]
      properties:
        userId: { type: string }
        wallet: { $ref: '#/components/schemas/ManagedWallet' }
    WalletChallengeResponse:
      type: object
      required: [transaction, networkPassphrase]
      properties:
        transaction: { type: string }
        networkPassphrase: { type: string }
    WalletVerifyRequest:
      type: object
      required: [transaction]
      properties:
        transaction: { type: string }
    WalletVerifyResponse:
      type: object
      required: [userId, address]
      properties:
        userId: { type: string }
        address: { type: string }
    ManagedWallet:
      type: object
      required: [userId, address, provisioned, network]
      properties:
        userId: { type: string }
        address: { type: string }
        provisioned: { type: boolean }
        network: { type: string, enum: [testnet, mainnet] }
    BuildTransactionRequest:
      type: object
      required: [userId, operations]
      properties:
        userId: { type: string }
        memo: { type: string }
        operations:
          type: array
          items: { type: object }
    BuildTransactionResponse:
      type: object
      required: [xdr, hash]
      properties:
        xdr: { type: string }
        hash: { type: string }
    SubmitTransactionRequest:
      type: object
      required: [xdr]
      properties:
        xdr: { type: string }
    SubmitTransactionResponse:
      type: object
      required: [hash, successful, horizonUrl]
      properties:
        hash: { type: string }
        successful: { type: boolean }
        ledger: { type: integer }
        horizonUrl: { type: string }
    ScoutTier:
      type: integer
      enum: [1, 2, 3]
      description: 'SCOUT reputation tier (multipliers 1.0 / 1.2 / 1.5)'
    SettlementDriverInput:
      type: object
      required: [userId, address, tier]
      properties:
        userId: { type: string }
        address: { type: string }
        tier: { $ref: '#/components/schemas/ScoutTier' }
    CreateSettlementBatchRequest:
      type: object
      required: [grossAmount, drivers]
      properties:
        grossAmount: { type: string }
        asset: { $ref: '#/components/schemas/AssetRef' }
        drivers:
          type: array
          items: { $ref: '#/components/schemas/SettlementDriverInput' }
    AssetRef:
      type: object
      required: [code]
      properties:
        code: { type: string }
        issuer: { type: string }
    SettlementSplit:
      type: object
      required: [authorities, driverRewards, treasury]
      properties:
        authorities: { type: string }
        driverRewards: { type: string }
        treasury: { type: string }
    SettlementDriverPayout:
      type: object
      required: [userId, address, tier, multiplier, amount]
      properties:
        userId: { type: string }
        address: { type: string }
        tier: { $ref: '#/components/schemas/ScoutTier' }
        multiplier: { type: number }
        amount: { type: string }
    SettlementBatch:
      type: object
      required: [id, createdAt, network, grossAmount, asset, split, driverPayouts, sourceAddress, authoritiesAddress, treasuryAddress, txHash, horizonUrl]
      properties:
        id: { type: string }
        createdAt: { type: string }
        network: { type: string, enum: [testnet, mainnet] }
        grossAmount: { type: string }
        asset: { $ref: '#/components/schemas/AssetRef' }
        split: { $ref: '#/components/schemas/SettlementSplit' }
        driverPayouts:
          type: array
          items: { $ref: '#/components/schemas/SettlementDriverPayout' }
        sourceAddress: { type: string }
        authoritiesAddress: { type: string }
        treasuryAddress: { type: string }
        txHash: { type: string }
        horizonUrl: { type: string }
    SettlementBatchPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/SettlementBatch' }
        nextCursor: { type: string, nullable: true }
    OffRampStatus:
      type: string
      enum: [pending_user_transfer_start, pending_anchor, completed, error]
    CreateOffRampWithdrawalRequest:
      type: object
      required: [amount]
      properties:
        amount: { type: string }
        asset: { $ref: '#/components/schemas/AssetRef' }
        fiatCurrency: { type: string }
        settlementBatchId: { type: string }
    OffRampSession:
      type: object
      required: [id, provider, sandbox, status, interactiveUrl, amount, asset, fiatCurrency, createdAt, updatedAt]
      properties:
        id: { type: string }
        provider: { type: string, enum: [mercuryo] }
        sandbox: { type: boolean }
        status: { $ref: '#/components/schemas/OffRampStatus' }
        interactiveUrl: { type: string }
        amount: { type: string }
        asset: { $ref: '#/components/schemas/AssetRef' }
        fiatCurrency: { type: string }
        fiatAmountEstimate: { type: string }
        anchorAccount: { type: string }
        settlementBatchId: { type: string }
        stellarTxHash: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    OffRampSessionPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/OffRampSession' }
        nextCursor: { type: string, nullable: true }